refactor(transactions): extract hooks + sub-components (FINLYNQ-111 phase 2)
Mechanical, behaviour-preserving extraction of the transactions god-component
(2063 -> 1021 lines, ~50% reduction). No UX / fetch-timing / state-shape change.

Hooks (src/app/(app)/transactions/_hooks/):
- useLookups        — accounts/categories/holdings mount triplet
- useTxColumnPrefs  — column layout + legacy localStorage migration + debounced PUT
- useTxSortPref     — header sort (cycle desc->asc->null) + debounced PUT
- useTxFilterPrefs  — per-column filters + debounced PUT
- useTransactions   — txns/total/loading + loadTxns (page owned by page-level state)

Components (src/app/(app)/transactions/_components/):
- TransactionTable      — header (sort/filter/drag) + body cell-render switch
- ColumnFilterPopover   — per-column filter popover (moved verbatim)
- SplitBadge            — split indicator badge

Shared types moved to ./_types.ts (Transaction/LinkedSibling/Account/Category/
Holding/ColFilterShape/SortPref/ColumnPref) so the page + extracted files share
one authority. Each hook moves the EXISTING state + effects verbatim (same deps,
fetch URLs, setState, debounce timings); cycleSort/setColFilter still reset the
page to 0 via an onChange callback. The two set-state-in-effect eslint findings
pre-date this change (verbatim-moved from the original page).
